requirements.txt
tqdm==4.11.2
opencv-python==3.4.0.12
Shapely==1.6.4.post1
python3-tk
matplotlib==1.5.3
numpy==1.14.2
tensorflow-gpu==1.4.0
Cython==0.24.1
ipython==5.1.0
Pillow

